feat!: implement stream value batching according to p3
In preview3 applications will be able to write a list of values of type
T to a stream<T> using a single method call - this functionality will be
used to batch the transmissions on the stream - eventually values may
also be buffered

+++ crates/transport-nats/src/lib.rs
@@ -212,7 +212,7 @@ impl ToSubject for &Subject {
 }
 
 impl wrpc_transport::Subject for Subject {
-    fn child(&self, i: Option<u32>) -> Self {
+    fn child(&self, i: Option<u64>) -> Self {
         if let Some(i) = i {
             Self(format!("{}.{i}", self.0).to_subject())
         } else {
